+++ packages/sdk/babel.config.cjs
@@ -5,4 +5,5 @@
 /** @type {import('ts-jest').JestConfigWithTsJest} */
 module.exports = {
     presets: ['module:metro-react-native-babel-preset'],
+    plugins: [['@babel/plugin-transform-private-methods', { loose: true }]],
 };
